backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters